ocr-server/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
easyocr==1.7.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
firebase-admin==6.5.0
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1